bookcover-crawler/requirements.txt
beautifulsoup4
pandas
Pillow
requests
tqdm
